+++ curator/validators/filters.py
@@ -1,7 +1,7 @@
 from voluptuous import *
-from ..defaults import settings
+from ..defaults import settings, filtertypes
 from ..exceptions import ConfigurationError
-from . import SchemaCheck, filtertypes
+from . import SchemaCheck
 import logging
 logger = logging.getLogger(__name__)
 
@@ -18,28 +18,7 @@ def filtertype():
 def structure():
     # This is to first ensure that only the possible keys/filter elements are
     # there, and get a dictionary back to work with.
-    retval = {
-        Optional('aliases'): Any(str, [str], unicode, [unicode]),
-        Optional('allocation_type'): Any(str, unicode),
-        Optional('count'): Coerce(int),
-        Optional('direction'): Any(str, unicode),
-        Optional('disk_space'): float,
-        Optional('epoch'): Any(Coerce(int), None),
-        Optional('exclude'): Any(int, str, unicode, bool, None),
-        Optional('field'): Any(str, unicode, None),
-        Optional('key'): Any(str, unicode),
-        Optional('kind'): Any(str, unicode),
-        Optional('max_num_segments'): Coerce(int),
-        Optional('reverse'): Any(int, str, unicode, bool, None),
-        Optional('source'): Any(str, unicode),
-        Optional('state'): Any(str, unicode),
-        Optional('stats_result'): Any(str, unicode, None),
-        Optional('timestring'): Any(str, unicode, None),
-        Optional('unit'): Any(str, unicode),
-        Optional('unit_count'): Coerce(int),
-        Optional('use_age'): Boolean(),
-        Optional('value'): Any(int, float, str, unicode, bool),
-    }
+    retval = settings.structural_filter_elements()
     retval.update(filtertype())
     return Schema(retval)